# backend/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "RSVP System",
  "version": "0.1.0"
 },
 "paths": {
...
  "/api/events/{event_id}/invitations": {
   "get": {
    "summary": "List Invitations",
    "operationId": "list_invitations_api_events__event_id__invitations_get",
    "parameters": [
     {
      "name": "event_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Event Id"
      }
     },
     {
      "name": "sort",
      "in": "query",
      "required": false,
      "schema": {
       "enum": [
        "name",
        "-name",
        "updated",
        "-updated",
        "status"
       ],
       "type": "string",
       "default": "name",
       "title": "Sort"
      }
     },
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 500,
       "minimum": 1,
       "default": 100,
       "title": "Limit"
      }
     },
     {
      "name": "offset",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "minimum": 0,
       "default": 0,
       "title": "Offset"
      }
     },
     {
      "name": "q",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Q"
      }
     },
     {
      "name": "rsvp_status",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "enum": [
          "no_response",
          "coming",
          "not_coming",
          "maybe"
         ],
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Rsvp Status"
      }
     },
     {
      "name": "group_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "title": "Group Id"
      }
     },
     {
      "name": "missing_phone",
      "in": "query",
      "required": false,
      "schema": {
       "type": "boolean",
       "default": false,
       "title": "Missing Phone"
      }
     },
     {
      "name": "invalid_phone",
      "in": "query",
      "required": false,
      "schema": {
       "type": "boolean",
       "default": false,
       "title": "Invalid Phone"
      }
     },
     {
      "name": "message_state",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Message State"
      }
     },
     {
      "name": "session",
      "in": "cookie",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Session"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/InvitationPage"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "post": {
    "summary": "Create Invitation",
    "operationId": "create_invitation_api_events__event_id__invitations_post",
    "parameters": [
     {
      "name": "event_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Event Id"
      }
     },
     {
      "name": "session",
      "in": "cookie",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Session"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/InvitationIn"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/InvitationOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/events/{event_id}/invitations/export": {
   "get": {
    "summary": "Export Invitations",
    "operationId": "export_invitations_api_events__event_id__invitations_export_get",
    "parameters": [
     {
      "name": "event_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Event Id"
      }
     },
     {
      "name": "lang",
      "in": "query",
      "required": false,
      "schema": {
       "enum": [
        "he",
        "en"
       ],
       "type": "string",
       "default": "he",
       "title": "Lang"
      }
     },
     {
      "name": "q",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Q"
      }
     },
     {
      "name": "rsvp_status",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "enum": [
          "no_response",
          "coming",
          "not_coming",
          "maybe"
         ],
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Rsvp Status"
      }
     },
     {
      "name": "group_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "title": "Group Id"
      }
     },
     {
      "name": "missing_phone",
      "in": "query",
      "required": false,
      "schema": {
       "type": "boolean",
       "default": false,
       "title": "Missing Phone"
      }
     },
     {
      "name": "invalid_phone",
      "in": "query",
      "required": false,
      "schema": {
       "type": "boolean",
       "default": false,
       "title": "Invalid Phone"
      }
     },
     {
      "name": "message_state",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Message State"
      }
     },
     {
      "name": "session",
      "in": "cookie",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Session"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/invitations/{invitation_id}": {
   "get": {
    "summary": "Get Invitation",
    "operationId": "get_invitation_api_invitations__invitation_id__get",
    "parameters": [
     {
      "name": "invitation_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Invitation Id"
      }
     },
     {
      "name": "session",
      "in": "cookie",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Session"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/InvitationOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "put": {
    "summary": "Update Invitation",
    "operationId": "update_invitation_api_invitations__invitation_id__put",
    "parameters": [
     {
      "name": "invitation_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Invitation Id"
      }
     },
     {
      "name": "session",
      "in": "cookie",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Session"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/InvitationIn"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/InvitationOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "delete": {
    "summary": "Delete Invitation",
    "operationId": "delete_invitation_api_invitations__invitation_id__delete",
    "parameters": [
     {
      "name": "invitation_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Invitation Id"
      }
     },
     {
      "name": "session",
      "in": "cookie",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Session"
      }
     }
    ],
    "responses": {
     "204": {
      "description": "Successful Response"
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/events/{event_id}/groups": {
   "get": {
    "summary": "List Groups",
    "operationId": "list_groups_api_events__event_id__groups_get",
    "parameters": [
     {
      "name": "event_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Event Id"
      }
     },
     {
      "name": "session",
      "in": "cookie",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Session"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/GroupOut"
         },
         "title": "Response List Groups Api Events  Event Id  Groups Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/groups/{group_id}": {
   "put": {
    "summary": "Rename Group",
    "operationId": "rename_group_api_groups__group_id__put",
    "parameters": [
     {
      "name": "group_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Group Id"
      }
     },
     {
      "name": "session",
      "in": "cookie",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Session"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/GroupIn"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/GroupOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "delete": {
    "summary": "Delete Group",
    "operationId": "delete_group_api_groups__group_id__delete",
    "parameters": [
     {
      "name": "group_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Group Id"
      }
     },
     {
      "name": "session",
      "in": "cookie",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Session"
      }
     }
    ],
    "responses": {
     "204": {
      "description": "Successful Response"
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "ContactIn": {
    "properties": {
     "phone": {
      "type": "string",
      "maxLength": 50,
      "minLength": 1,
      "title": "Phone"
     },
     "member_index": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Member Index"
     }
    },
    "type": "object",
    "required": [
     "phone"
    ],
    "title": "ContactIn"
   },
   "ContactOut": {
    "properties": {
     "id": {
      "type": "integer",
      "title": "Id"
     },
     "member_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Member Id"
     },
     "phone_raw": {
      "type": "string",
      "title": "Phone Raw"
     },
     "phone_e164": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Phone E164"
     },
     "phone_valid": {
      "type": "boolean",
      "title": "Phone Valid"
     }
    },
    "type": "object",
    "required": [
     "id",
     "member_id",
     "phone_raw",
     "phone_e164",
     "phone_valid"
    ],
    "title": "ContactOut"
   },
...
   "GroupIn": {
    "properties": {
     "name": {
      "type": "string",
      "maxLength": 100,
      "minLength": 1,
      "title": "Name"
     }
    },
    "type": "object",
    "required": [
     "name"
    ],
    "title": "GroupIn"
   },
   "GroupOut": {
    "properties": {
     "id": {
      "type": "integer",
      "title": "Id"
     },
     "name": {
      "type": "string",
      "title": "Name"
     }
    },
    "type": "object",
    "required": [
     "id",
     "name"
    ],
    "title": "GroupOut"
   },
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
   "InvitationIn": {
    "properties": {
     "display_name": {
      "type": "string",
      "maxLength": 200,
      "minLength": 1,
      "title": "Display Name"
     },
     "group_name": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 100
       },
       {
        "type": "null"
       }
      ],
      "title": "Group Name"
     },
     "notes": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Notes"
     },
     "members": {
      "items": {
       "$ref": "#/components/schemas/MemberIn"
      },
      "type": "array",
      "title": "Members",
      "default": []
     },
     "contacts": {
      "items": {
       "$ref": "#/components/schemas/ContactIn"
      },
      "type": "array",
      "title": "Contacts",
      "default": []
     }
    },
    "type": "object",
    "required": [
     "display_name"
    ],
    "title": "InvitationIn"
   },
   "InvitationOut": {
    "properties": {
     "id": {
      "type": "integer",
      "title": "Id"
     },
     "event_id": {
      "type": "integer",
      "title": "Event Id"
     },
     "display_name": {
      "type": "string",
      "title": "Display Name"
     },
     "group_name": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Group Name"
     },
     "rsvp_token": {
      "type": "string",
      "title": "Rsvp Token"
     },
     "rsvp_status": {
      "type": "string",
      "enum": [
       "no_response",
       "coming",
       "not_coming",
       "maybe"
      ],
      "title": "Rsvp Status"
     },
     "attendee_count": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Attendee Count"
     },
     "rsvp_updated_at": {
      "anyOf": [
       {
        "type": "string",
        "format": "date-time"
       },
       {
        "type": "null"
       }
      ],
      "title": "Rsvp Updated At"
     },
     "rsvp_source": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Rsvp Source"
     },
     "notes": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Notes"
     },
     "message_state": {
      "type": "string",
      "title": "Message State"
     },
     "members": {
      "items": {
       "$ref": "#/components/schemas/MemberOut"
      },
      "type": "array",
      "title": "Members"
     },
     "contacts": {
      "items": {
       "$ref": "#/components/schemas/ContactOut"
      },
      "type": "array",
      "title": "Contacts"
     }
    },
    "type": "object",
    "required": [
     "id",
     "event_id",
     "display_name",
     "group_name",
     "rsvp_token",
     "rsvp_status",
     "attendee_count",
     "rsvp_updated_at",
     "rsvp_source",
     "notes",
     "message_state",
     "members",
     "contacts"
    ],
    "title": "InvitationOut"
   },
   "InvitationPage": {
    "properties": {
     "items": {
      "items": {
       "$ref": "#/components/schemas/InvitationOut"
      },
      "type": "array",
      "title": "Items"
     },
     "total": {
      "type": "integer",
      "title": "Total"
     }
    },
    "type": "object",
    "required": [
     "items",
     "total"
    ],
    "title": "InvitationPage"
   },
...
   "MemberIn": {
    "properties": {
     "name": {
      "type": "string",
      "maxLength": 200,
      "minLength": 1,
      "title": "Name"
     }
    },
    "type": "object",
    "required": [
     "name"
    ],
    "title": "MemberIn"
   },
   "MemberOut": {
    "properties": {
     "id": {
      "type": "integer",
      "title": "Id"
     },
     "name": {
      "type": "string",
      "title": "Name"
     }
    },
    "type": "object",
    "required": [
     "id",
     "name"
    ],
    "title": "MemberOut"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   }
  }
 }
}